requirements.txt
pyotp
flask
gunicorn
PyYAML
requests
